move chart types into types folder
