+++ src/components/Chart/ChartComposed.tsx
@@ -16,11 +16,11 @@ import {
 import { Props } from 'recharts/types/component/DefaultLegendContent';
 
 import { formatTimestamp, getColors } from 'helpers';
+import { ChartComposedProps, ChartConfigType } from 'types';
 
 import { ChartTooltip } from './ChartTooltip';
 import { formatYAxis } from './helpers/formatYAxis';
 import { StartEndTick } from './helpers/StartEndTick';
-import { ChartComposedProps, ChartConfigType } from './helpers/types';
 import { useChartComposedData } from './hooks/useChartComposedData';
 
 export const ChartComposed = ({
